+++ src/api/health.go
@@ -18,24 +18,22 @@ type BlockchainMetadata struct {
 
 // HealthResponse is returned by the /health endpoint
 type HealthResponse struct {
-	BlockchainMetadata            BlockchainMetadata `json:"blockchain"`
-	Version                       readable.BuildInfo `json:"version"`
-	CoinName                      string             `json:"coin"`
-	DaemonUserAgent               string             `json:"user_agent"`
-	OpenConnections               int                `json:"open_connections"`
-	OutgoingConnections           int                `json:"outgoing_connections"`
-	IncomingConnections           int                `json:"incoming_connections"`
-	Uptime                        wh.Duration        `json:"uptime"`
-	CSRFEnabled                   bool               `json:"csrf_enabled"`
-	CSPEnabled                    bool               `json:"csp_enabled"`
-	WalletAPIEnabled              bool               `json:"wallet_api_enabled"`
-	GUIEnabled                    bool               `json:"gui_enabled"`
-	UnversionedAPIEnabled         bool               `json:"unversioned_api_enabled"`
-	JSON20RPCEnabled              bool               `json:"json_rpc_enabled"`
-	UserBurnFactor                uint32             `json:"user_burn_factor"`
-	UnconfirmedBurnFactor         uint32             `json:"unconfirmed_burn_factor"`
-	UserMaxTransactionSize        uint32             `json:"user_max_transaction_size"`
-	UnconfirmedMaxTransactionSize uint32             `json:"unconfirmed_max_transaction_size"`
+	BlockchainMetadata    BlockchainMetadata `json:"blockchain"`
+	Version               readable.BuildInfo `json:"version"`
+	CoinName              string             `json:"coin"`
+	DaemonUserAgent       string             `json:"user_agent"`
+	OpenConnections       int                `json:"open_connections"`
+	OutgoingConnections   int                `json:"outgoing_connections"`
+	IncomingConnections   int                `json:"incoming_connections"`
+	Uptime                wh.Duration        `json:"uptime"`
+	CSRFEnabled           bool               `json:"csrf_enabled"`
+	CSPEnabled            bool               `json:"csp_enabled"`
+	WalletAPIEnabled      bool               `json:"wallet_api_enabled"`
+	GUIEnabled            bool               `json:"gui_enabled"`
+	UnversionedAPIEnabled bool               `json:"unversioned_api_enabled"`
+	JSON20RPCEnabled      bool               `json:"json_rpc_enabled"`
+	UserVerifyTxn         readable.VerifyTxn `json:"user_verify_transaction"`
+	UnconfirmedVerifyTxn  readable.VerifyTxn `json:"unconfirmed_verify_transaction"`
 }
 
 // healthHandler returns node health data
@@ -71,23 +69,21 @@ func healthHandler(c muxConfig, csrfStore *CSRFStore, gateway Gatewayer) http.Ha
 				BlockchainMetadata: readable.NewBlockchainMetadata(health.BlockchainMetadata),
 				TimeSinceLastBlock: wh.FromDuration(timeSinceLastBlock),
 			},
-			Version:                       c.health.BuildInfo,
-			CoinName:                      c.health.CoinName,
-			DaemonUserAgent:               userAgent,
-			OpenConnections:               health.OutgoingConnections + health.IncomingConnections,
-			OutgoingConnections:           health.OutgoingConnections,
-			IncomingConnections:           health.IncomingConnections,
-			Uptime:                        wh.FromDuration(health.Uptime),
-			CSRFEnabled:                   csrfStore.Enabled,
-			CSPEnabled:                    !c.disableCSP,
-			UnversionedAPIEnabled:         c.enableUnversionedAPI,
-			GUIEnabled:                    c.enableGUI,
-			JSON20RPCEnabled:              c.enableJSON20RPC,
-			WalletAPIEnabled:              walletAPIEnabled,
-			UserBurnFactor:                params.UserBurnFactor,
-			UserMaxTransactionSize:        params.UserMaxTransactionSize,
-			UnconfirmedBurnFactor:         health.UnconfirmedBurnFactor,
-			UnconfirmedMaxTransactionSize: health.UnconfirmedMaxTransactionSize,
+			Version:               c.health.BuildInfo,
+			CoinName:              c.health.CoinName,
+			DaemonUserAgent:       userAgent,
+			OpenConnections:       health.OutgoingConnections + health.IncomingConnections,
+			OutgoingConnections:   health.OutgoingConnections,
+			IncomingConnections:   health.IncomingConnections,
+			Uptime:                wh.FromDuration(health.Uptime),
+			CSRFEnabled:           csrfStore.Enabled,
+			CSPEnabled:            !c.disableCSP,
+			UnversionedAPIEnabled: c.enableUnversionedAPI,
+			GUIEnabled:            c.enableGUI,
+			JSON20RPCEnabled:      c.enableJSON20RPC,
+			WalletAPIEnabled:      walletAPIEnabled,
+			UserVerifyTxn:         readable.NewVerifyTxn(params.UserVerifyTxn),
+			UnconfirmedVerifyTxn:  readable.NewVerifyTxn(health.UnconfirmedVerifyTxn),
 		})
 	}
 }